home and register page design done
